+++ src/components/Feed/Feed.tsx
@@ -1,19 +1,63 @@
 import React, { useState, useEffect } from 'react';
 import Card from '../Card/Card';
 import './Feed.css';
-import { CleanArticle } from '../../apiCalls'
+import { CleanArticle, fetchNewsData, getSentiment } from '../../apiCalls'
 
 interface PropsFeed {
-  articles: CleanArticle[]
+  userSentiment: number | null
 }
 
-const Feed = (): JSX.Element => {
+const Feed = ({ userSentiment }: PropsFeed): JSX.Element => {
 
   const [ articles, setArticles ] = useState<CleanArticle[]>([])
+  const [ error, setError ] = useState('')
+
+  const getSentimentScores = (cleanArticles: CleanArticle[]): Promise<number[]> => {
+    return Promise.all(
+      cleanArticles.map((article: CleanArticle) => {
+        return getSentiment(article.abstract)
+      })
+    )
+  }
+
+  useEffect((): void => {
+    fetchNewsData()
+      .then((cleanArticles: CleanArticle[]): void => {
+        getSentimentScores(cleanArticles)
+          .then((response: number[]) => {
+
+            const scoredArticles = cleanArticles.map((article, i) => {
+               article.sentiment = response[i] || 0
+               return article;
+            })
+            sortBySentiment(userSentiment, scoredArticles)
+          })
+      })
+      .catch(error => setError(error.message))
+  }, [])
+
+  const sortBySentiment = (newSentiment: number|null, scoredArticles: CleanArticle []): void => {
+    let sortedArticles : CleanArticle[] = [];
+
+    if (newSentiment === -1) {
+      sortedArticles = scoredArticles.sort((articleA, articleB) => {
+        return articleB.sentiment - articleA.sentiment;
+      })
+    } else if (newSentiment === 1) {
+      sortedArticles = scoredArticles.sort((articleA, articleB) => {
+        return articleA.sentiment - articleB.sentiment;
+      })
+    } else {
+      sortedArticles = scoredArticles.sort((articleA, articleB) => 0.5 - Math.random());
+    }
+
+    setArticles(sortedArticles)
+  }
 
   return (
     <div className="articles-container">
       <section className="articles-display">
+      { !articles.length && <h1>Loading</h1> }
         {articles.map(article =>
           <Card
           title={article.title}
@@ -21,6 +65,7 @@ const Feed = (): JSX.Element => {
           key={article.title}
           />
         )}
+        {error && <h2>{error}</h2>}
       </section>
     </div>
   )
